Added Patch method for Patients resources Added various validation attributes on DTOs Refactored IBrowsable to specify mutliple links for one resource Changes the link header template to match the RFC5988.

+++ src/MedEasy.DTO/BodyWeightInfo.cs
@@ -14,8 +14,8 @@ namespace MedEasy.DTO
         /// <summary>
         /// Value of the measure
         /// </summary>
-        [Range(0, float.MaxValue)]
-        public float Value { get; set; }
+        [Range(0, double.MaxValue)]
+        public decimal Value { get; set; }
 
     }
 }
